Changed the type of the zear to long because we get the pubdate of an book as year month day

+++ app/models/Book.java
@@ -45,7 +45,7 @@ public class Book extends Model
     
     @Constraints.Required
     @Formats.NonEmpty
-    public int year;
+    public long year;
     
     public boolean tradeable;
